tests and fixtures improvements: - persistence - channel - memory - move tests to separate file - domain - re-export fixtures from the underlying modules of channel, validator & asset
